chore: move create to commands module
