+++ cli/src/commands/mod.rs
@@ -1,0 +1,3 @@
+mod create;
+
+pub use create::*;
