+++ cli/src/utils/mod.rs
@@ -1,3 +1,5 @@
+mod capture;
 mod is_valid_extension;
 
+pub use capture::*;
 pub use is_valid_extension::*;
